Makes tool refuse to use insecure HTTP
Ffs will now refuse to use anything other than http by default, as this
would transmit credentials over the network in plain text. This can be
overridden by using the --force option for the terminally masochistic.

Also improves the error discrimination at the top level `main` function
in order to stop the error handler printing these same credentials back
to the console in the event of a failed HTTP request.