Allows user to specify local timezone
Sort-of addresses issue #1 (the possibility of getting the search range wrong
because the timezone on the JIRA server doesn't match the local timesone on
the user's computer). Ideally we could just specify the timezone in the
search range but JIRA won't let you do that. It will, however, honour
the timezone that a user has configured on their account when filtering
search results.

This changeset Aat least lets users manually sync things up if their system
timezone is different to the timezone configured on their jira account.